+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/TilesetOptions.h
@@ -1,11 +1,14 @@
 #pragma once
 
 #include "Cesium3DTilesSelection/Library.h"
+#include <memory>
 #include <optional>
 #include <string>
 #include <vector>
 
 namespace Cesium3DTilesSelection {
+
+class ITileExcluder;
 
 /**
  * @brief Options for configuring the parsing of a {@link Tileset}'s content
@@ -19,6 +22,18 @@ struct CESIUM3DTILESSELECTION_API TilesetContentOptions {
    * water mask extension.
    */
   bool enableWaterMask = false;
+
+  /**
+   * @brief Whether to generate smooth normals when normals are missing in the
+   * original Gltf.
+   *
+   * According to the Gltf spec: "When normals are not specified, client
+   * implementations should calculate flat normals." However, calculating flat
+   * normals requires duplicating vertices. This option allows the gltfs to be
+   * sent with explicit smooth normals when the original gltf was missing
+   * normals.
+   */
+  bool generateMissingNormalsSmooth = false;
 };
 
 /**
@@ -175,6 +190,13 @@ struct CESIUM3DTILESSELECTION_API TilesetOptions {
    * types.
    */
   bool renderTilesUnderCamera = true;
+
+  /**
+   * @brief A list of interfaces that are given an opportunity to exclude tiles
+   * from loading and rendering. If any of the excluders indicate that a tile
+   * should not be loaded, it will not be loaded.
+   */
+  std::vector<std::shared_ptr<ITileExcluder>> excluders;
 
   /**
    * @brief Options for configuring the parsing of a {@link Tileset}'s content