+++ Cesium3DTilesSelection/src/Tile.cpp
@@ -9,6 +9,7 @@
 #include "CesiumGeometry/Axis.h"
 #include "CesiumGeometry/AxisTransforms.h"
 #include "CesiumGeospatial/Transforms.h"
+#include "CesiumGltf/Model.h"
 #include "CesiumUtility/Tracing.h"
 #include "TileUtilities.h"
 #include "upsampleGltfForRasterOverlays.h"
@@ -266,7 +267,10 @@ void Tile::loadContent() {
            gltfUpAxis,
            pPrepareRendererResources =
                tileset.getExternals().pPrepareRendererResources,
-           pLogger = tileset.getExternals().pLogger](
+           pLogger = tileset.getExternals().pLogger,
+           generateMissingNormalsSmooth =
+               tileset.getOptions()
+                   .contentOptions.generateMissingNormalsSmooth](
               std::shared_ptr<IAssetRequest>&& pRequest) mutable {
             CESIUM_TRACE("loadContent worker thread");
             const IAssetResponse* pResponse = pRequest->response();
@@ -329,6 +333,12 @@ void Tile::loadContent() {
                     transform,
                     boundingVolume,
                     projections);
+
+                pContent->rasterOverlayProjections = std::move(projections);
+
+                if (generateMissingNormalsSmooth) {
+                  pContent->model->generateMissingNormalsSmooth();
+                }
 
                 if (pPrepareRendererResources) {
                   CESIUM_TRACE("prepareInLoadThread");
@@ -706,7 +716,24 @@ void Tile::update(
             pLoadingTile->getOverlay().getTileProvider();
 
         // Try to replace this placeholder with real tiles.
-        if (pProvider && !pProvider->isPlaceholder()) {
+        if (pProvider && !pProvider->isPlaceholder() && this->getContent()) {
+          // Find a suitable projection in the content.
+          // If there isn't one, the mesh doesn't have the right texture
+          // coordinates for this overlay and we need to kick it back to the
+          // unloaded state to fix that.
+          // In the future, we could add the ability to add the required
+          // texture coordinates without starting over from scratch.
+          const auto& projections =
+              this->getContent()->rasterOverlayProjections;
+          auto it = std::find(
+              projections.begin(),
+              projections.end(),
+              pProvider->getProjection());
+          if (it == projections.end()) {
+            this->unloadContent();
+            return;
+          }
+
           this->_rasterTiles.erase(
               this->_rasterTiles.begin() +
               static_cast<
@@ -718,7 +745,9 @@ void Tile::update(
               projectRectangleSimple(pProvider->getProjection(), *pRectangle);
           CesiumUtility::IntrusivePointer<RasterOverlayTile> pTile =
               pProvider->getTile(projectedRectangle, this->getGeometricError());
-          this->_rasterTiles.emplace_back(pTile);
+
+          RasterMappedTo3DTile& mapped = this->_rasterTiles.emplace_back(pTile);
+          mapped.setTextureCoordinateID(int32_t(it - projections.begin()));
         }
 
         continue;
@@ -849,22 +878,24 @@ void Tile::upsampleParent(
   pTileset->getAsyncSystem()
       .runInWorkerThread([&parentModel,
                           transform = this->getTransform(),
-                          projections,
+                          projections = std::move(projections),
                           pSubdividedParentID,
                           boundingVolume = this->getBoundingVolume(),
                           pPrepareRendererResources =
                               pTileset->getExternals()
-                                  .pPrepareRendererResources]() {
+                                  .pPrepareRendererResources]() mutable {
         std::unique_ptr<TileContentLoadResult> pContent =
             std::make_unique<TileContentLoadResult>();
         pContent->model =
             upsampleGltfForRasterOverlays(parentModel, *pSubdividedParentID);
+
         if (pContent->model) {
           pContent->updatedBoundingVolume = Tile::generateTextureCoordinates(
               pContent->model.value(),
               transform,
               boundingVolume,
               projections);
+          pContent->rasterOverlayProjections = std::move(projections);
         }
 
         void* pRendererResources = nullptr;