+++ Cesium3DTilesSelection/src/TileUtilities.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "Cesium3DTilesSelection/BoundingVolume.h"
-#include "Cesium3DTilesSelection/CartographicSelection.h"
+#include "CesiumGeospatial/CartographicPolygon.h"
 #include "CesiumGeospatial/GlobeRectangle.h"
 #include <vector>
 
@@ -35,22 +35,24 @@ const CesiumGeospatial::GlobeRectangle* obtainGlobeRectangle(
  * @brief Returns whether the tile is completely inside a polygon.
  *
  * @param boundingVolume The {@link Cesium3DTilesSelection::BoundingVolume} of the tile.
- * @param cartographicSelections The list of polygon selections to check.
+ * @param cartographicPolygons The list of polygons to check.
  * @return Whether the tile is completely inside a polygon.
  */
 bool withinPolygons(
     const BoundingVolume& boundingVolume,
-    const std::vector<CartographicSelection>& cartographicSelections) noexcept;
+    const std::vector<CesiumGeospatial::CartographicPolygon>&
+        cartographicPolygons) noexcept;
 
 /**
  * @brief Returns whether the tile is completely inside a polygon.
  *
  * @param rectangle The {@link CesiumGeospatial::GlobeRectangle} of the tile.
- * @param cartographicSelections The list of polygon selections to check.
+ * @param cartographicPolygons The list of polygons to check.
  * @return Whether the tile is completely inside a polygon.
  */
 bool withinPolygons(
     const CesiumGeospatial::GlobeRectangle& rectangle,
-    const std::vector<CartographicSelection>& cartographicSelections) noexcept;
+    const std::vector<CesiumGeospatial::CartographicPolygon>&
+        cartographicPolygons) noexcept;
 } // namespace Impl
 } // namespace Cesium3DTilesSelection