+++ CesiumGeospatial/src/CartographicPolygon.cpp
@@ -1,11 +1,9 @@
 
-#include "Cesium3DTilesSelection/CartographicSelection.h"
+#include "CesiumGeospatial/CartographicPolygon.h"
 #include <array>
 #include <mapbox/earcut.hpp>
 
-using namespace CesiumGeospatial;
-
-namespace Cesium3DTilesSelection {
+namespace CesiumGeospatial {
 
 static std::vector<uint32_t>
 triangulatePolygon(const std::vector<glm::dvec2>& polygon) {
@@ -33,7 +31,7 @@ triangulatePolygon(const std::vector<glm::dvec2>& polygon) {
 
     // check if the difference crosses the antipole
     if (glm::abs(point[0]) > CesiumUtility::Math::ONE_PI) {
-      if (point[0] > 0) {
+      if (point[0] > 0.0) {
         point[0] -= CesiumUtility::Math::TWO_PI;
       } else {
         point[0] += CesiumUtility::Math::TWO_PI;
@@ -41,7 +39,7 @@ triangulatePolygon(const std::vector<glm::dvec2>& polygon) {
     }
   }
 
-  localPolygons.push_back(localPolygon);
+  localPolygons.emplace_back(std::move(localPolygon));
 
   indices = mapbox::earcut<uint32_t>(localPolygons);
   return indices;
@@ -76,11 +74,11 @@ computeBoundingRectangle(const std::vector<glm::dvec2>& polygon) {
     // check if the difference crosses the antipole
     if (glm::abs(dif_west) > CesiumUtility::Math::ONE_PI) {
       // east wrapping past the antipole to the west
-      if (dif_west > 0) {
+      if (dif_west > 0.0) {
         west = point1.x;
       }
     } else {
-      if (dif_west < 0) {
+      if (dif_west < 0.0) {
         west = point1.x;
       }
     }
@@ -89,11 +87,11 @@ computeBoundingRectangle(const std::vector<glm::dvec2>& polygon) {
     // check if the difference crosses the antipole
     if (glm::abs(dif_east) > CesiumUtility::Math::ONE_PI) {
       // west wrapping past the antipole to the east
-      if (dif_east < 0) {
+      if (dif_east < 0.0) {
         east = point1.x;
       }
     } else {
-      if (dif_east > 0) {
+      if (dif_east > 0.0) {
         east = point1.x;
       }
     }
@@ -102,14 +100,9 @@ computeBoundingRectangle(const std::vector<glm::dvec2>& polygon) {
   return CesiumGeospatial::GlobeRectangle(west, south, east, north);
 }
 
-CartographicSelection::CartographicSelection(
-    const std::string& targetTextureName,
-    const std::vector<glm::dvec2>& polygon,
-    bool isForCulling)
-    : _targetTextureName(targetTextureName),
-      _vertices(polygon),
+CartographicPolygon::CartographicPolygon(const std::vector<glm::dvec2>& polygon)
+    : _vertices(polygon),
       _indices(triangulatePolygon(polygon)),
-      _boundingRectangle(computeBoundingRectangle(polygon)),
-      _isForCulling(isForCulling) {}
+      _boundingRectangle(computeBoundingRectangle(polygon)) {}
 
-} // namespace Cesium3DTilesSelection
+} // namespace CesiumGeospatial