+++ CHANGES.md
@@ -1,9 +1,17 @@
 # Change Log
 
-### Next Release - ?
+### v0.7.1 - ????
+
+##### Fixes :wrench:
+
+- Fixed a bug introduced in v0.7.0 where Bing credits were not being collected.
+
+### v0.7.0 - 2021-09-01
 
 ##### Breaking Changes :mega:
 
+- Renamed the `Cesium3DTiles` namespace and library to `Cesium3DTilesSelection`.
+- Deleted `Cesium3DTilesSelection::Gltf` and moved functionality into `CesiumGltf::Model`.
 - Renamed `Rectangle::intersect` and `GlobeRectangle::intersect` to `computeIntersection`.
 - `RasterOverlay` and derived classes now require a `name` parameter to their constructors.
 - Changed the type of texture coordinate IDs used in the raster overlay system from `uint32_t` to `int32_t`.
@@ -14,6 +22,7 @@
 
 - Added `Future<T>::isReady`.
 - Added `Future<T>::share`, which returns a `SharedFuture<T>` and allows multiple continuations to be attached.
+- Added an option in `TilesetOptions::ContentOptions` to generate smooth normals when the original glTFs were missing normals.
 - Added `ImageManipulation` class to `CesiumGltfReader`.
 - Added `Math::roundUp` and `Math::roundDown`.
 - Added `Rectangle::computeUnion`.